Can now push opposing player.
Bug Fix: Food not allowing pushforce scaling. Minor typo.

Modified push force handling, pushforce is now a percentage so it scales far slower.

Need to put movement caps so pushing an opponent doesnt cause you to move backwards or cause you to move faster than normal

Notably this push doesnt work if the opposing player is pushing back on you. This may be a good mechanic to keep, so for now it will be treated as a feature not a bug