Changed Loop Behavior on Size change.
The loop was not properly setup causing infinite loops. This has been corrected.

After fixing the loop the resizing does push the sumo out of range. However, it does push the sumo more than expected, not sure why at this time. However it seems like it may work well as a feature rather than a bug.

Also, might try to put "resize" as a script that is called to ensure it's logic is consistent for size up and size down.